requirements.txt
python-dotenv==0.13.*
python-telegram-bot==12.*
requests==2.23.*



